@@ -4,6 +4,7 @@
 	"homepage": "https://emil.pw",
 	"scripts": {
 		"build": "nx run web:build:production",
+		"build:web": "nx run web:export",
 		"dev": "nx run web:serve",
 		"lint": "nx run-many --target=lint --all",
 		"postinstall": "husky install"
@@ -12,46 +13,49 @@
 	"dependencies": {
 		"@reduxjs/toolkit": "1.8.2",
 		"async-mutex": "0.3.2",
-		"core-js": "3.6.5",
+		"core-js": "^3.6.5",
 		"date-fns": "2.28.0",
 		"nanoid": "3.3.3",
+		"next": "13.0.0",
 		"query-string": "7.1.1",
-		"react": "18.0.0",
-		"react-aui": "^1.8.12",
-		"react-dom": "18.0.0",
+		"react": "18.2.0",
+		"react-aui": "^1.9.1",
+		"react-dom": "18.2.0",
 		"react-ga4": "1.4.1",
 		"react-helmet-async": "1.3.0",
-		"react-is": "18.0.0",
+		"react-is": "18.2.0",
 		"react-redux": "8.0.2",
 		"react-router-dom": "6.3.0",
 		"redux-logger": "3.0.6",
 		"redux-persist": "6.0.0",
 		"reflect-metadata": "0.1.13",
 		"regenerator-runtime": "0.13.7",
-		"styled-components": "5.3.5",
-		"tslib": "2.3.0",
+		"styled-components": "5.3.6",
+		"tslib": "^2.3.0",
 		"uuid": "8.3.2"
 	},
 	"devDependencies": {
+		"@babel/preset-react": "^7.14.5",
 		"@commitlint/cli": "17.2.0",
 		"@commitlint/config-conventional": "17.2.0",
 		"@nrwl/cli": "15.0.0",
 		"@nrwl/eslint-plugin-nx": "15.0.0",
-		"@nrwl/jest": "15.0.0",
+		"@nrwl/jest": "15.3.0",
 		"@nrwl/linter": "15.0.0",
+		"@nrwl/next": "^15.3.0",
 		"@nrwl/nx-cloud": "14.7.0",
-		"@nrwl/react": "15.0.0",
-		"@nrwl/web": "15.0.0",
+		"@nrwl/react": "15.3.0",
+		"@nrwl/web": "15.3.0",
 		"@nrwl/workspace": "15.0.0",
 		"@semantic-release/changelog": "6.0.1",
 		"@semantic-release/git": "10.0.1",
 		"@semantic-release/gitlab": "9.5.0",
-		"@testing-library/react": "13.1.1",
+		"@testing-library/react": "13.4.0",
 		"@trivago/prettier-plugin-sort-imports": "3.2.0",
-		"@types/jest": "27.4.1",
-		"@types/node": "18.11.9",
-		"@types/react": "18.0.20",
-		"@types/react-dom": "18.0.6",
+		"@types/jest": "28.1.1",
+		"@types/node": "16.11.7",
+		"@types/react": "18.0.25",
+		"@types/react-dom": "18.0.9",
 		"@types/react-helmet": "6.1.5",
 		"@types/react-is": "17.0.3",
 		"@types/react-router-dom": "5.3.3",
@@ -60,24 +64,28 @@
 		"@types/uuid": "8.3.4",
 		"@typescript-eslint/eslint-plugin": "~5.18.0",
 		"@typescript-eslint/parser": "~5.18.0",
-		"babel-jest": "27.5.1",
+		"babel-jest": "28.1.1",
 		"babel-plugin-styled-components": "1.10.7",
 		"compression-webpack-plugin": "10.0.0",
 		"eslint": "~8.12.0",
+		"eslint-config-next": "13.0.0",
 		"eslint-config-prettier": "8.1.0",
+		"eslint-plugin-cypress": "^2.10.3",
 		"eslint-plugin-import": "2.26.0",
 		"eslint-plugin-jsx-a11y": "6.6.1",
-		"eslint-plugin-react": "7.31.8",
+		"eslint-plugin-react": "7.31.11",
 		"eslint-plugin-react-hooks": "4.6.0",
 		"eslint-plugin-unused-imports": "2.0.0",
 		"husky": "8.0.2",
-		"jest": "27.5.1",
-		"jest-environment-jsdom": "27.5.1",
+		"jest": "28.1.1",
+		"jest-environment-jsdom": "28.1.1",
+		"next-transpile-modules": "^10.0.0",
 		"nx": "15.0.0",
 		"prettier": "2.5.1",
-		"react-test-renderer": "18.0.0",
+		"react-test-renderer": "18.2.0",
 		"semantic-release": "19.0.5",
-		"ts-jest": "27.1.4",
+		"ts-jest": "28.0.5",
+		"ts-node": "10.9.1",
 		"typescript": "~4.6.2",
 		"webpack-bundle-analyzer": "^4.6.1"
 	},
