@@ -1,9 +1,10 @@
-module.exports = {
+/* eslint-disable */
+export default {
 	displayName: 'web',
-	preset: '../../jest.preset.ts',
+	preset: '../../jest.preset.js',
 	transform: {
 		'^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nrwl/react/plugins/jest',
-		'^.+\\.[tj]sx?$': 'babel-jest',
+		'^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nrwl/next/babel'] }],
 	},
 	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
 	coverageDirectory: '../../coverage/apps/web',
